+++ app/src/main/java/org/simple/clinic/monthlyReports/questionnaire/component/BaseComponent.kt
@@ -1,0 +1,10 @@
+package org.simple.clinic.monthlyReports.questionnaire.component
+
+import android.os.Parcelable
+import com.squareup.moshi.JsonClass
+import kotlinx.parcelize.Parcelize
+
+@JsonClass(generateAdapter = true)
+@Parcelize
+open class BaseComponent : Parcelable
+