+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 ### Internal
 
 - Update adding a new screen document
-- Added deleted_at column to `Questionnaire`s table
 - Migrate `BloodSugarUnitSelectionDialog` to use view effects
+- Add deleted_at column to `Questionnaire`s table
+- Add Questionnaire repository
 
 ### Internal